+++ packages/presentation/src/PresentationTool.tsx
@@ -15,7 +15,6 @@ import {BoundaryElementProvider, Flex} from '@sanity/ui'
 import {
   lazy,
   type ReactElement,
-  startTransition,
   Suspense,
   useCallback,
   useEffect,
@@ -164,7 +163,15 @@ export default function PresentationTool(props: {
         },
       )
     }
-  }, [params.perspective, state.perspective, navigate, state.viewport, params.viewport])
+  }, [
+    params.perspective,
+    state.perspective,
+    navigate,
+    state.viewport,
+    params.viewport,
+    params.rev,
+    params.prefersLatestPublished,
+  ])
 
   const [overlaysConnection, setOverlaysConnection] = useState<ChannelStatus>('connecting')
   const [loadersConnection, setLoadersConnection] = useState<ChannelStatus>('connecting')
@@ -192,6 +199,12 @@ export default function PresentationTool(props: {
     }
   }, [channel, popups, popups.size])
 
+  // This workaround can be removed once the need to set prefersLatestPublished is no longer there
+  const perspectiveRef = useRef(state.perspective)
+  useEffect(() => {
+    perspectiveRef.current = state.perspective
+  }, [state.perspective])
+
   useEffect(() => {
     const target = iframeRef.current?.contentWindow
 
@@ -212,11 +225,19 @@ export default function PresentationTool(props: {
           onStatusUpdate: setOverlaysConnection,
           onEvent(type, data) {
             if ((type === 'visual-editing/focus' || type === 'overlay/focus') && 'id' in data) {
-              navigate({
-                type: data.type,
-                id: data.id,
-                path: data.path,
-              })
+              navigate(
+                {
+                  type: data.type,
+                  id: data.id,
+                  path: data.path,
+                },
+                {
+                  prefersLatestPublished:
+                    'isDraft' in data || perspectiveRef.current === 'previewDrafts'
+                      ? undefined
+                      : 'true',
+                },
+              )
             } else if (type === 'visual-editing/navigate' || type === 'overlay/navigate') {
               const {title, url} = data
               if (frameStateRef.current.url !== url) {
@@ -320,30 +341,28 @@ export default function PresentationTool(props: {
   useEffect(() => {
     const interval = setInterval(
       () =>
-        startTransition(() =>
-          setLiveQueries((liveQueries) => {
-            if (Object.keys(liveQueries).length < 1) {
-              return liveQueries
-            }
+        setLiveQueries((liveQueries) => {
+          if (Object.keys(liveQueries).length < 1) {
+            return liveQueries
+          }
 
-            const now = Date.now()
-            const hasAnyExpired = Object.values(liveQueries).some(
-              (liveQuery) =>
-                liveQuery.heartbeat !== false && now > liveQuery.receivedAt + liveQuery.heartbeat,
-            )
-            if (!hasAnyExpired) {
-              return liveQueries
+          const now = Date.now()
+          const hasAnyExpired = Object.values(liveQueries).some(
+            (liveQuery) =>
+              liveQuery.heartbeat !== false && now > liveQuery.receivedAt + liveQuery.heartbeat,
+          )
+          if (!hasAnyExpired) {
+            return liveQueries
+          }
+          const next = {} as LiveQueriesState
+          for (const [key, value] of Object.entries(liveQueries)) {
+            if (value.heartbeat !== false && now > value.receivedAt + value.heartbeat) {
+              continue
             }
-            const next = {} as LiveQueriesState
-            for (const [key, value] of Object.entries(liveQueries)) {
-              if (value.heartbeat !== false && now > value.receivedAt + value.heartbeat) {
-                continue
-              }
-              next[key] = value
-            }
-            return next
-          }),
-        ),
+            next[key] = value
+          }
+          return next
+        }),
       MIN_LOADER_QUERY_LISTEN_HEARTBEAT_INTERVAL,
     )
     return () => clearInterval(interval)
@@ -393,6 +412,18 @@ export default function PresentationTool(props: {
       channel?.send('overlays', 'presentation/blur', undefined)
     }
   }, [channel, params.id, params.path])
+
+  // Handle opening the published document when previewing published
+  useEffect(() => {
+    if (
+      state.perspective === 'published' &&
+      params.id &&
+      !params.rev &&
+      !params.prefersLatestPublished
+    ) {
+      navigate({}, {prefersLatestPublished: 'true'})
+    }
+  }, [navigate, params.id, params.prefersLatestPublished, params.rev, state.perspective])
 
   // Dispatch a navigation message when the preview param changes
   useEffect(() => {