+++ packages/presentation/src/components/ErrorCard.tsx
@@ -1,5 +1,8 @@
 import {Box, Button, Card, type CardProps, Container, Flex, Inline, Stack, Text} from '@sanity/ui'
 import type {ReactElement, ReactNode} from 'react'
+import {useTranslation} from 'sanity'
+
+import {presentationLocaleNamespace} from '../i18n'
 
 export function ErrorCard(
   props: {
@@ -11,14 +14,18 @@ export function ErrorCard(
 ): ReactElement {
   const {children, message, onRetry, onContinueAnyway, ...restProps} = props
 
-  const retryButton = <Button fontSize={1} mode="ghost" onClick={onRetry} text="Retry" />
+  const {t} = useTranslation(presentationLocaleNamespace)
+
+  const retryButton = (
+    <Button fontSize={1} mode="ghost" onClick={onRetry} text={t('error-card.retry-button.text')} />
+  )
   const continueAnywayButton = (
     <Button
       fontSize={1}
       mode="ghost"
       tone="critical"
       onClick={onContinueAnyway}
-      text="Continue anyway"
+      text={t('error-card.continue-button.text')}
     />
   )
 
@@ -29,7 +36,7 @@ export function ErrorCard(
           <Stack space={4}>
             <Stack space={3}>
               <Text size={1} weight="semibold">
-                An error occured
+                {t('error-card.title')}
               </Text>
               <Text muted size={1}>
                 {message}