+++ packages/overlays/CHANGELOG.md
@@ -72,6 +72,15 @@ All notable changes to this project will be documented in this file. See
   * dependencies
     * @sanity/visual-editing bumped from 1.7.0 to 1.7.1
 
+## [2.3.44](https://github.com/sanity-io/visual-editing/compare/overlays-v2.3.43...overlays-v2.3.44) (2024-05-05)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @sanity/visual-editing bumped from 1.8.17 to 1.8.18
+
 ## [2.3.43](https://github.com/sanity-io/visual-editing/compare/overlays-v2.3.42...overlays-v2.3.43) (2024-05-02)