+++ packages/preview-kit-compat/CHANGELOG.md
@@ -36,6 +36,20 @@ All notable changes to this project will be documented in this file. See
   * devDependencies
     * @sanity/visual-editing-helpers bumped from 0.6.0 to 0.6.1
 
+## [1.4.21](https://github.com/sanity-io/visual-editing/compare/preview-kit-compat-v1.4.20...preview-kit-compat-v1.4.21) (2024-05-05)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @sanity/client to v6.17.2 ([#1449](https://github.com/sanity-io/visual-editing/issues/1449)) ([928b45c](https://github.com/sanity-io/visual-editing/commit/928b45c47c66c5e1ddbe8b00783efd399b9fa17a))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * @repo/visual-editing-helpers bumped from 0.6.9 to 0.6.10
+
 ## [1.4.20](https://github.com/sanity-io/visual-editing/compare/preview-kit-compat-v1.4.19...preview-kit-compat-v1.4.20) (2024-04-26)